implement getLogs in sqlite indexer frontend
