+++ packages/store-sync/src/index.ts
@@ -3,3 +3,5 @@ export * from "./createStoreSync";
 export * from "./SyncStep";
 export * from "./isTableRegistrationLog";
 export * from "./logToTable";
+export * from "./tablesWithRecordsToLogs";
+export * from "./tableToLog";
